+++ CHANGELOG.md
@@ -4,15 +4,15 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [2.2.0-alpha.0] - 2018-05-04
+## [2.2.0-preview.1] - 2018-05-09
 ### Impulse Module and More
 - New Cinemachine Impulse module for event-driven camera shakes
-- New Event Helper script CinemachineEventMapper maps UnityEvents to Virtual Cameras
 - New Event Helper script CinemachineTriggerAction takes action on Collider and Collider2D enter/exit events, and exposes them as UnityEvents
 - Added Focal Length or Named FOV presets for Camera Lens
 - Added support for Physical Camera
 - Collider: now returns TargetIsObscured if the target is offscreen (great for cameras with fixed aim)
 - Collider: added Minimum Occlusion Time setting, to ignore fleeting obstructions
+- Collider: added Transparent Layers mask, to specify solid objects that don't obstruct view
 - Collider: damping will no longer take the camera through obstacles
 - NoiseSettings: added checkbox for pure sine-wave instead of Perlin wave
 - Added NoiseSettings editor with signal preview